.github/pyproject.toml
[project]
name = "ghpages"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jinja2>=3.1.6",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "ruff>=0.14.13",
    "ty>=0.0.12",
    "djlint>=1.36.4",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.djlint]
max_attribute_length = 100
max_line_length = 120
ignore = "H023"
